TrafficCapture/dockerSolution/src/main/docker/otelcol/requirements.txt
backoff==1.11.1
certifi==2023.7.22
charset-normalizer==3.3.0
Deprecated==1.2.14
googleapis-common-protos==1.61.0
grpcio==1.59.0
idna==3.4
importlib-metadata==6.8.0
opentelemetry-api==1.20.0
opentelemetry-distro==0.41b0
opentelemetry-exporter-otlp==1.12.0rc1
opentelemetry-exporter-otlp-proto-grpc==1.12.0rc1
opentelemetry-exporter-otlp-proto-http==1.12.0rc1
opentelemetry-instrumentation==0.41b0
opentelemetry-proto==1.12.0rc1
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
protobuf==3.20.2
requests==2.31.0
typing_extensions==4.8.0
urllib3==2.0.7
wrapt==1.15.0
zipp==3.17.0
